requirements.txt
asgiref==3.7.2
Django==5.0.2
sqlparse==0.4.4
tzdata==2024.1
psycopg[binary]==3.1.12
djangorestframework==3.14.0